requirements.txt
streamlit
pandas
youtube-transcript-api
requests
nltk
spacy
openai
python-dotenv
wordcloud